Updated mongoose to 6.16
Upgrade mongoose webserver (mongoose.c and mongoose.h) to 6.16 [1].

The issue found in SWUpdate with current mongoose is that the ping/pong
mechanismus in Websocket does not work when the IPC blocks due to I/O.
It is required to write just a part of the data in non blockin mode, but
the webserver requires that the handler consumes all data. New version
allows to return the bytes still available in the buffer and the server
will call the handler again.

[1] dce60c6dbb096f3b96e1a45cbfdfd55e18b38bb6.

Signed-off-by: Stefano Babic <sbabic@denx.de>