mongoose: set receive limit to 256KB
After 6.132, mongoose introduces with commit
93286836f944cf71c0f344eab8c1b0a29666f82a that all data are read from
socket during receive poll. This can exhaust memory and cause a
deadlock, because the transfer to IPC is much slower. Avoid to read full
data and sync the receive with IPC by setting a limit for the incoming
data to be quite the same as the size of the UDS in the IPC.

Signed-off-by: Stefano Babic <sbabic@denx.de>