Webserver: drop multiplication factor for monitoring
Commit 37089b0 introduces monitoring of the data connection when an
update is started. The value passed with the "-t" parameter is then
interpreted as timer value to check the connection, while a broken
connection is recognized only if no data is received after the timer
elapses at least two times. Drop the multiplication factor and check for
incoming data when the timer elapses.

Signed-off-by: Stefano Babic <sbabic@denx.de>