suricatta: add IPC copmmand to enable / disable
Add commands to "trigger" a check by the server and to enable / disable
the polling mechanismus. This can be useful on target with slow and
expensive connection to the Internet, and it is preferable that the end
customer decides when to check for an update.

An additional IPC command is added - it accepts a json command in the
format
	"enable" : "true"
		or
	"enable" : "false"

Signed-off-by: Stefano Babic <sbabic@denx.de>