suricatta: add enable / disable at startup
For systems without a fixed network connection, it makes sense that the
connection to a backend server is done on demand. Command line for
suricatta can contain -e (enable, default) or -d (disable). If disabled,
suricatta does not try to connect to the backend server when the polling
timer elapses. Behavior can be changed at runtime without restarting
SWUpdate via IPC.

Signed-off-by: Stefano Babic <sbabic@denx.de>