Intercepts the output of shell scripts
Output of Shell scripts called by shell script's handler or as pre- and post
install scripts is lost because it does not go into the common tracing
facility in SWUpdate. SWUpdate checks the error code from the scripts,
but important information to debug an error cause is simply lost.

Rework how shell scripts are called to get the output:

- a process is created for the script
- stdout and stderr are redirected to a pipe
- SWUpdate (parent process) reads the pipe, and use TRACE and ERROR to
store output and errors from the scripts.

Signed-off-by: Stefano Babic <sbabic@denx.de>