suricatta: added option to bind to interface/IP address.
This patch adds the -f cmdline arg and the "interface" parameter to suricatta
to optionally bind communication channel to a specific interface or IP address.
To do so, a new config option is added to curl channel interface, which then
exploits CURLOPT_INTERFACE option.
Binding to an interface is useful in multi-interface devices, where routing
can be policy-based (e.g with output traffic classification rules based on
src address/device).

Signed-off-by: Luca Pesce <luca.pesce@vimar.com>
Signed-off-by: Pedro Aguilar <pedro.aguilar@vimar.com>