stream_interface: Keep reading the cpio padding
If any, up to 512 bytes from the socket until the client stops writing.

This avoids that the streaming client gets an EPIPE error when trying to write
the padding to the IPC socket. This can happen when the streaming client is
downloading very slowly and swupdate, after reading the trailer, closes the pipe.
The cpio utility always adds a 512 bytes alignment after the file trailer,
independently of the given format.

Signed-off-by: Pedro Aguilar <pedro.aguilar@vimar.com>
Signed-off-by: Luca Pesce <luca.pesce@vimar.com>