hawkbit: send log of the update to the server
No log is sent to the server - in case of failure, the reson is stored
on the device and cannot be seen on the server. This makes difficult to
identify the cause of failure. Add a temporary thread that collects all
traces and errors and sends them to the Hawkbit server. The thread
exists for the time a SWU is installed, and then exits gracefully.

Signed-off-by: Stefano Babic <sbabic@denx.de>