Select CHANNEL_CURL_SSL with SURICATTA_SSL
Suricatta's SSL does not depend on OpenSSL directly
but on curl with SSL support and on the SSL_IMPL.
Hence, depend on the SSL_IMPL and select the CHANNEL_CURL_SSL
config on activating it.

There are ifdefs where both of the two symbols are checked.
Let them only check CHANNEL_CURL_SSL.

Signed-off-by: Bastian Germann <bastiangermann@fishpost.de>
Acked-by: Stefano Babic <sbabic@denx.de>