BUG: all images runs in streaming mode
Commit 16f44d8b4 is supposed to run all partioners before installing
the first image in streaming mode. The loop is intended to run for all
partitioners mode, but without checking the type, it runs for all
images, independently if it is supposed to change a partition. This
results in an update failure if the image must be treated by a handler
that does not support streaming mode.

Signed-off-by: Stefano Babic <sbabic@denx.de>