progress: make -r reboot only if something was installed
swupdate-progress may be used as a convenient tool to monitor updates
happening in the background. It also provides an option to reboot the
system when the installation completes. SWUpdate may send a "SUCCESS"
status even if no artifacts were installed (possiby because all images
had install-if-higher and the current system already running the latest
and greatest version of the OS). Avoid a reboot by checking cur_step
from the status update messages to determine what to do on "SUCCESS".

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>