configs: remove debian_defconfig
The debian_defconfig is not used as-is by its two main users:
swupdate's debian branch which has one change on top of master
("debian: temporary drop swuforwarder") and the Debian project's package
which replaces the config with its own config file.

Logically, the file belongs to the debianization, so remove it from the
undebianized project.

Signed-off-by: Bastian Germann <bastiangermann@fishpost.de>
Acked-by: Stefano Babic <sbabic@denx.de>