Add option to disable cpio CRC check
Disable CRC check in cpio header if sha256 is enabled. CRC in CPIO is not
a real crc, but it is simply the sum of all bytes belonging to a file as
32 bit value. It is very weak and does not add any further safety if sha256
is activated. CPIO tool is buggy on Linux distros and CRC is not computed
correctly and set to zero when a file is larger than 2GB.
While cpio on OE is patched to fix this, a SWU built outside OE has a wrong
computed CRC and SWUpdate will abort the update.

This option drops the check when the stronger sha256 check is activated.

Signed-off-by: Stefano Babic <sbabic@denx.de>