swupdate-progress: make sure bar array is 0-terminated
The bar filling was based on the assumption that there are sizeof(bar) - 1
characters, but terminating 0 was never set. The bar array was 60 chars,
and the fprintf formatting string expected to find 60 chars, thus printing one
garbage char, and the missing 0-terminator was causing the app to crash.

The commit increases the size of bar by 1, and 0-terminates it.

Signed-off-by: Tomas Frydrych <tomas@tomtain.com>