suricatta: add IPC copmmand to trigger an update
This extends commit e705a76a2 and add a command to force a check on the
server without waiting for a polling time. This is useful in case
the device is mostly offline, and when it is online, it should check and
update as soon as possible without waiting.

An additional IPC command is added - it accepts a json command in the
format:
            "trigger" : "true"
                    or
            "trigger : "false"

trigger is one shot - after checking, trigger is reset automatically by
the daemon. The external process is in charge to send multiple trigger
requests if desired.

Signed-off-by: Stefano Babic <sbabic@denx.de>