lua: Prevent mtdname and path from clobbering each other
The lua properties "mtdname" and "path" are both initialized to the path
property in struct img_type, and both values were written back to the
path property when update struct img_type from the lua table. This means
that which ever one was encountered last when iterating would be the one
assigned to actual path property. This in particular causes problems
when attempting to chain to another handler that uses the path property
from lua, for example the following lua code could randomly fail if
mtdname was parsed last when chaining handlers:

 function my_handler(image)
   image.path = "/foo/bar"
   swupdate.call_handler("archive", image)
 end

To resolve this, use a dedicated variable for the "mtdname" property and
update the appropriate handlers to use it instead of the "path" property

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Reviewed-by: Stefano Babic <sbabic@denx.de>