swupdate-progress: Terminate bar array string
The bar array contains the progress bar. Its length is 60 characters
but space for a terminating \0 was forgotten. This causes problems when
feeding bar to fprintf().

By introducing the constant bar_len there is now a single place to
change in case a different progress bar length is required.

The original code only used 59 characters for printing the progress bar
but the fprintf() call reserved a width of 60 characters. The code now
uses 60 visible characters for the progress bar.

Suggested-by: Tomas Frydrych <tomas@tomtain.com>
Signed-off-by: Mark Jonas <toertel@gmail.com>