Regression: allow SW version comparisons as strings
Semantic version are now allowed, but each version is set to be compared
as semantic version. If the string does not match the requirements,
version is artificially set to 0. This breaks compatibility with past
because simple string comparison with any char was possible.

Reassemble the order of the comparison as:

- if versions are in format x.y.t.z, then compare them
- if versions are suitable for semantic version, compare them
- else simple use strcmp to return if they match

Signed-off-by: Stefano Babic <sbabic@denx.de>
Signed-off-by: Stijn Devriendt <sde@unmatched.eu>
Tested-by: Heiko Schocher <hs@denx.de>