diskpart: fix crash in case if GPT
Switch to the same behavior as with MBR and rewrite completely the
partition table instead of trying to adjust a single entry.
The most reliable way to set new partitions seems to be to drop the old
table and to write from scratch the new table, without trying to adjust
previous entries. This is the way done for MBR, and this patch aligns
GPT in the same way.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Heiko Schocher <hs@denx.de>
Tested-by: Heiko Schocher <hs@denx.de>