Configure colors for stdout / stderr
SWUpdate uses colors to log on the stdout / sdterr. This lets to
identify better each loglevel, but the default values can conflict with
the setup of the terminal and makes the output difficult to read. Add a
section in the configuration file and let configure the colors for each
loglevel.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Claudius Heine <ch@denx.de>