encrypted images: fix one block sized image decryption
SWUpdate has a bug in the decryption handling. For encrypted image
content that is exactly one AES block (16 bytes) of size (from an
unencrypted file of < 16 bytes), the output is empty.

The reason for this effect is the first swupdate_DECRYPT_update writing
0 bytes and the copyfile() implementation skipping the final decryption
step because it checks on the written bytes.

Check swupdate_DECRYPT_update return for zero, and call
swupdate_DECRYPT_final() directly afterwards.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Tested-by: Stefano Babic <sbabic@denx.de>