Makefile.flags: Add missing wolfSSL definitions
The openssl compatibility layer is located in the openssl subdirectory of
the wolfssl include dir. OPENSSL_ALL is needed to enable some of the
compatibility layer's optional definitions.

WOLFSSL_APACHE_HTTPD is needed to circumvent a wolfSSL memory bug (?)
that is triggered if OPENSSL_VERSION_NUMBER is < 0x10100000L.
WOLFSSL_APACHE_HTTPD makes it raise the number. With image encryption
enabled and this undefined the bug can be triggered by running `make test`.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>