pkcs11: Implement based on wolfSSL and p11-kit
Add a PKCS#11 implementation for encrypted images. This replaces the
image decryption once activated. The interface stays the same but the key
is interpreted as PKCS#11 URI, which is first parsed via p11-kit. This
seems to be the most advanced parser implementation out there.

For the PKCS#11 abstraction, wolfSSL is used because it has first-class
support and AES is supported well, as opposed to OpenSSL that needs an
engine loaded at run time with the most prominent PKCS#11 engine not
supporting AES.

wolfSSL comes up with the concept of a crypto device that you register so
that the normal AES operations can be outsourced to them.

wolfSSL does not implement PKCS#7 padding in the AES operations, which is
needed because the encrypted images use it.  Implement the unpadding in
swupdate_DECRYPT_final which operates on one block that is remembered from
the last AES decryption.

The decryption and PKCS#11 handling need context structs to live during
all decryption steps. Extend the existing decryption_key struct with them.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Acked-by: Torben Hohn <torben.hohn@linutronix.de>