handlers: ubivol: add property decompressed-size
Add the property decompressed-size that works similar to the decrypted-size
property.

This was posted to the mailing list by Matt over two years ago
("Update fails when ubifs is compressed - fixed and patch attached"),
so I assume it is meant to be released under the SWUpdate license.
The first version did not use properties and was rejected.

Co-developed-by: Matt <mattbeaumontgts@gmail.com>
[Add commit msg. Change decompressed_size to decompressed-size]
Signed-off-by: Bastian Germann <bage@linutronix.de>