Rework emit download progress
There is a notification to signal how many bytes are downloaded, but
this was later introduced and does not match SWUpdate's design. In fact,
because it is not called by the main process, it uses the notifier to
send messages to the main process. The notification caused in the past
flood of data to the installer, and does not fill the foreseen fields in
the progress IPC to forward download statistics.

This adds a swupdate_download_update() that can be called by any
process. If it is called by the main process, emit a progress message to
the listeners, else it prepares a message and sends it as notification
to the progress notifier. A new enum for the status (RECOVERY_DWL) is
added to let the progress notifier to identify the type of message.

Signed-off-by: Stefano Babic <sbabic@denx.de>