Move global structure for installer in private header
The "struct installer" is a shared global data structure used to store
setup for the current installation and to lock if an update is running
by storing the current status. It is loaded just by the initiator of an
update (the installers as streaming or file). It should be better moved
to a private file instead of beeing in util.h that is included by all
due to get logging facilities.

Signed-off-by: Stefano Babic <sbabic@denx.de>