IPC: drop unused parameter and change name
The "info" buffer used to trigger an update is quite unused. The name of
the original SWU is pushed in case of local update, but it is not used.
Drop the filename and change the name to a more generic data, as this
will be used to pass a request structure with several fields instead of
pure formatted strings.

Signed-off-by: Stefano Babic <sbabic@denx.de>