IPC: rework install IPC for future extensions
The IPC was originally very simple and used just to send a SWU to
SWUpdate. Later, additional features were added to communicate with
internal SWUpdate's processes, like suricatta or the installer. The same
message for install was (mis)used, letting the interface quite confused.
Adding new commands is then difficult, because changes in the messages
mean changes in more subsystems.This patch addresses all these topics.

The message used to communicate with processes is renamed to "procmsg"
from "instmsg", but without changing its fields. A new "instmsg" is
created from scratch and this contains an "install request" structuire,
that the originator must fill with setup for the installation. Each
interface is updated to the new interface. The API ro external processes
is changed, but it will contain a field to track the API version and to
allow to have compatibility layers in future to avoid further breakages.
The "install request" is thought to add more information as in the past,
and to allow to change an installation per instance, while most
parameters are now set just at the startup via configuration file or
command line parameters.

Signed-off-by: Stefano Babic <sbabic@denx.de>