IPC: let decide selections at runtime
It is useful to change the selection type at runtime if more as one
section inside sw-description can be installed. This activates a
selection sent via IPC to be used for the next update. The default
configuration is cached and will be active at next update if the request
does not contain any valid selection.

Signed-off-by: Stefano Babic <sbabic@denx.de>