syslog: suppress message not belonging to log
Some messages sent to the notifier are not for logging, but they are for
the progress interface or they must reach a process. Currently, such as
messages are not recognized and a FATAL_UNKNOWN mesagge is sent to the
syslog server. This simply ignores messages that do not need to be
looged.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Christian Storm <christian.storm@siemens.com>