dry-run option overwritten by IPC request
An IPC request can contain if the install should run in dry-run mode,
but this overwrites the same option issued with command line parameter.
Store --dry-run into SWUpdate's structure and use it if the request does
not have set any value for dry-run. A request should indicate if dry-run
is set or if the default value (from --dry-run) should be taken.

Signed-off-by: Stefano Babic <sbabic@denx.de>