libconfig parser: fix incorrect type cast
The function `get_value_libconfig` contains an incorrect type cast
when parsing booleans which results in overriding other memory.

The bug can be triggered by having an sw-description like this:

```
bootloader_transaction_marker = false;

[...]

foo = {
    bar: {
        [...]
    }
};
```

and using SWUpdate's selection mechanism (`-e foo,bar`).

The selection mechanism will *not* work when sw-description contains
`bootloader_transaction_marker`.

The reason is that the value of `bootloader_transaction_marker` is
stored in the struct `swupdate_cfg` which contains the fields:

```
    bool bootloader_transaction_marker;
    char software_set[SWUPDATE_GENERAL_STRING_SIZE];
```

The incorrect type cast (`int` instead of `bool`) results in overriding
the value stored in the `software_set` and consequently breaks the
software selection mechanism.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>
Reviewed-by: Stefano Babic <sbabic@denx.de>