Drop separate installation path for local file
There are currently two separate paths for installing, via network
thread (used by OTA / IPC library) and from a local file. This increases
efforts during regression tests because both paths must be tested. In
case of local path, no network thread is created and file is scanned via
seek(), adding an implicit "installed-directly" to all artifacts.
Nevertheless, it happened to find dicrepancies between OTA and local
file update.
This patch is a big cleanup and drops the path from local file. The
network thread is always generated, andcode from swupdate-client is
integrated into SWUpdate to use the generic update path.
Specific functions duplicated for local path are dropped, and it is not
requested to forward in all steps if an updated is coming from local
file or from OTA.

These changes are thought to be compatible with the past, and updating
via -i (local file) is still supported. Indeed, the following different
behaviors are introduced:

- SWU check via -c is the same as dry-run mode. Not just sw-description
is parsed, but the whole update is checked without installing. This
extends the check's functionality and verifies the whole SWU.
- before these changes, -i results in an implicit "installed-drirectly"
for all artifacts. This is not maintained, and users should add the
attribute to own sw-description to avoid temporary copies on tmpdir.

Signed-off-by: Stefano Babic <sbabic@denx.de>