Drop unneeded wrapper save_state_string
This function is used just in one place to set the transaction flag. It
was thought to be generic, but the state is already saved using
set_/get_state, and this function is just a wrapper for bootloader
accessors, doing also some dangerous type conversion. This patch cleans
this up and calls directly the functions to access the bootloader's
environment.

Signed-off-by: Stefano Babic <sbabic@denx.de>