parser: split find_node function
The find_node() function searches for a path inside sw-description,
looking at different root start point (selection, board, or initial root.)
The caller gets a pointer to the element if found. In case of links, the caller
requires the whole path to the element to find a relative path from the found
element where the link is set. Split the function in two parts, the core
returns also the full path (as ** char array) where the element is found.

Signed-off-by: Stefano Babic <sbabic@denx.de>