parser: follow the links
Support for "links" is incomplete: a link can be set to a whole group
but it cannot be set to an element of a group because it is not
followed. In more complicate update scenarios, several updates are
provided where artifacts are grouped together.

The patch adds support for links in each element, so the following will
be supported:

	selection : {
	   images: (
		{
                    ref = "#./../../somecommon/images";
                },
		{
                    ref = "#./../../somespecific/images";
                },
                {
                    filename = "somefilename";
                    device = "/dev/somedevice";
                    type = "raw";
                }
	    )

Signed-off-by: Stefano Babic <sbabic@denx.de>