Enable tracing for early start up code
The current behavior is to set the log level in the *second*
getopt_long(3) parsing stage of SWUpdate. In other words, all log
messages with a finer granularity (e.g. INFO) which are created *before*
reaching the second getopt parsing stage are dropped since the default
log level is ERROR. This makes it (unneccesarily) difficult to debug
issues in the swupdate.cfg parsing code.
To allow for config overloading, we continue to parse the loglevel a
second time after loading and parsing swupdate.cfg.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>