Restructure swupdate data
The internal SWUpdate database was split to have a "globals"
sub-structure with the original goal to have here the parameters read
from the "globals" section in swupdate.cfg. However, entries were
already mixed and there is no real use case for it. Move most parameters
into the main structure, and create a structure for paramteres that can
be changed on each instance of an update (mainly via IPC). This avoids
to save somewhere the default parameters as it was done previously, and
if additional parameters can be changed in future via IPC, it is enough
to move them into this internal structure to let them automatically
restored after each run.

Signed-off-by: Stefano Babic <sbabic@denx.de>