Avoid re-opening and reading configuration file
During a regular startup, the configuration file (swupdate.cfg) is at
least open(2)ed and read(2) four times.

This patch introduces a new API for reading the configuration file,
which passes a handle to an in-memory copy of the configuration file
around.

For security reasons, the in-memory copy is not shared with
subprocesses, because a less privileged subprocess (setuid(3p)) might
not be allowed to read the configuration file. Future work could
optimize this further and check if the subprocess runs under the same
privileges and if so, share the configuration data via an appropriate
mechanism (e.g. mmap(3p).

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>