build system: do not rebuild if nothing changed
Try to avoid rebuilding things needlessly by:
 - removing FORCE from direct output files, trust make dependencies
to do the right thing.

 - make cfg-sanity-check a real file that depends on .config,
so the check can be reexecuted on config change but not otherwise.

 - make tools create tools/built-in.o: that file is expected to exist
and used as timestamp comparison, but nothing creates it currently.
That one is quite ugly as trying to create the file directly would
update it needlessly so an extra "pseudo-phony" target was added.

 - remove KBUILD_STR: the # character was incorrectly truncating
commands in $(cmd_$@), making Kbuild think the command line changed
(can be tested with KBUILD_VERBOSE=2) ; in our case filling simple
quotes should always work for version number, so skip the macro
altogether.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>