grubenv: fix buffer overflow on strtok function
grubenv file is a fixed block of memory (1024) not null-terminated.
Using strtok on this buffer can have unpredicted results based on
the memory content immediately after the buffer.

During an intensive SWUPDATE test, we noticed some random crashes
with the following error:

[ERROR] : SWUPDATE failed [0] ERROR bootloader/grub.c : grubenv_write : 174 : Not enough free space in envblk file, 1031
[ERROR] : SWUPDATE failed [0] ERROR corelib/installer.c : update_bootloader_env : 191 : Bootloader-specific error -1 updating its environment Software updated failed

It turned out to be related to strtok overrunning the env buffer.
The bottom of the grubenv file is filled with "#" characters, and
the latest cycle on the do_while loop was detecting the "####" string as
the `key` env token (very long string). This file is not terminated with
'\0' so strtok was not stopping until it went beyound the buffer area.
On a lucky run, the memory immediately after the buffer was filled with
0, so strtok returned a very long string for the `key` string, and a
null for `value`.

Sometimes the memory after the buffer was filled with some other data,
which made the strtok call go on finding a valid `value` string outside
the buffer, resulting in the swupdate crash. Since strtok is destructive
on the string it uses (replaces the string separator with '\0') this
could also lead to memory corruption.

With this patch, we allocate an extra byte for the '\0' character on the
buffer, to make sure strtok works as expected.

Signed-off-by: Elvis Kapllaj <e.kapllaj@asem.it>