Format extX filesystem
The diskpart handler can repartition a device but it was not able to
format with a filesystem. Commit c35590 introduces the format capability
for VFAT filesystem. This patch goes in the same direction and adds
support for extX (ext2 / ext3 / ext4) filesystem.

The code to initialize the filesystem is largely taken from the mke2fs
utility present on all Linux distribution. Code was rearranged to work
as library, and it was simplified using the default setup. It does not
support all features of mke2fs, and it substitutes the profiles used in
the original code with a fix set of features. It has the same
functionalities as :

	mke2fs -t [ext2 | ext3 | ext4] <device>

Signed-off-by: Stefano Babic <sbabic@denx.de>