Add optional CMS single signer verification
This change introduces a Kconfig parameter allowing CMS verification
when additional unrecognized signatures are included in the CMS stream.
Content verification is required against all signatures, but swupdate
only requires a single signature in the set to be verified against the
public key specified to swupdate. This operation requires manual
checking of signatures outside of the CMS_verify operation as OpenSSL
requires all signatures within the CMS envelop to verify.

Signed-off-by: Andrew Mulbrook <andrew.mulbrook@garmin.com>