util: Extend get_root to find LUKS devices
This helps in case of encrypted filesystem or device mapper.
The returned device read from partitions is usually a dm-X device and
this does not show which is the block device that contains it. Look in
sysfs and check if the device has "slaves" entries, indicating the
presence of an underlying device. If found, return this instead of the
device returned parsing /proc/partitions.

Signed-off-by: Stefano Babic <sbabic@denx.de>