read_lines_notify: fix segfault if a program prints a line starting with \0
if a program prints something starting with \0, string_split will
successfully allocate a string array of one element which is null
(so nlines == 0 on the next line)

There might be other problem with string length (we know how many
bytes we read), so instead of trying to work around it in all cases
just replace all nul bytes with @ like we often see in e.g. less.

This can be reproduced by executing a shell script that does
e.g. dd if=/dev/zero bs=1 count=1

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>