Improve error handling around archive_read_next_header
ARCHIVE_OK is the most likely result of the method, so make that code path the fastest.

archive_read_next_header may return ARCHIVE_WARN, which is not fatal and extracting can continue.
In particular, the message "Pathname can't be converted from UTF-8 to current locale" is of the
warning category only. The pathname will just be interpreted as a string of bytes, which is in
most cases just fine.

In case of error, get rid of the extra "1" appended to the error message.

This fixes that SWUpdate aborts when for example "ca-certificates" is added to the image inside the
tar file. Some of the filenames in that package have non-ASCII characters.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Acked-by: Stefano Babic <sbabic@denx.de>