pkcs11: Reuse one session for all operations
Currently, each PKCS#11 operation is run in its own session. Explicitly
open a session, so that it gets reused in the subsequent operations.
It will be closed with the already used wc_Pkcs11Token_Final.

This will trigger a conceptual bug in wolfSSL's PKCS#11 implementation for
wolfSSL versions < 4.8.0 but will improve decryption performance a lot.

Link: https://github.com/wolfSSL/wolfssl/issues/4143
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Holger Dengler <dengler@linutronix.de>
Acked-by: Stefano Babic <sbabic@denx.de>