mongoose: not block the sender if update is ended
The interface between the webserver and the installer was set in not
blocking mode. However, holding the data to be sent should be done just
in case SWUpdate reports -EBUSY, that is the pipe is full. In case
SWUpdate has already finished, it closes the pipe and returns -EPIPE.
The sender (mongoose) should then ignored it and consume incoming data
to unblock the sender.

Signed-off-by: Stefano Babic <sbabic@denx.de>