handlers: add diskformat handler
Currently the discpart handler only creates a new file system when the
partition table has changed.

In cases where there exists an unused and unformatted spare partion which
shall be activated by a later update, it is necessary to format such a
partition even if the partition table has not been altered.

Also there are cases when creating a file system on a single partition is
desired, but there is no partitioning scheme specified inside the
sw-description.

Therefore, this commit introduces a new diskformat handler which allows
creating a file system independently from the partitioning process.
Before creating a new file system it checks whether the desired file system
already exists and only writes if not.

In case it is desired to create a new file system anyway, this can be
achieved by setting the property "force" to "true".

	partitions: (
	{
		type = "diskformat";
		device = "/dev/mmcblk0p1";

		properties: {
			fstype = "vfat";
			force = "true";
		}
	})

The diskformat handler will be active when the configuration option
CONFIG_DISKFORMAT_HANDLER is selected.

Signed-off-by: Roland Gaudig <roland.gaudig@weidmueller.com>