Kconfig: use DISKFORMAT as flag
The following warnings are generated if DISKFORMAT_HANDLER is set
without setting DISKPART:

scripts/kconfig/conf  --silentoldconfig Kconfig
warning: (DISKFORMAT_HANDLER) selects DISKFORMAT which has unmet direct dependencies (DISKPART)
warning: (DISKFORMAT_HANDLER) selects DISKFORMAT which has unmet direct dependencies (DISKPART)

This is due to a dependency between the handlers, that is not necessary
because it is possible to link just one of them. Replace
CONFIG_DISKFORMAT with CONFIG_DISKPART_FORMAT in diskpüart handler, and
use CONFIG_DISKFORMAT as a flag to enable/disable the activation of the
filesystems creators.

Signed-off-by: Stefano Babic <sbabic@denx.de>