doc: Fix documentation for nolock/noinuse to match implementation.
The documentation for these flags does not appear to match the
implementation, by default these force flags should be "false"
not "true" since they should only be used in cases where they
are actually needed.

Since we always attempt the lock when nolock is set we need
to fix the docs to reflect that we are bypassing a lock attempt
error and not the lock attempt itself.

For noinuse we also do unconditionally check if the file
system is in use for logging purposes, reword this so that we
don't imply that the check is not done at all, only that the
results of the check will be ignored if set.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>