delta: add process to download chunks
The delta handler must download the missing parts of an artifacts.
Because it runs with high rights, the download itself should be done by
a separate process to avoid to break privilege separation. The process
gets userid and group id from the configuration file, and as fallback
will run with the same rights as the installer.

The downloader will wait for requests, and it will write the downloaded
data into the IPC pipe. Only one process is allowed to talk to the
downloader at once. If more as one user will exist in the future, the
access to the downloader should be queued (not done at the present).

Signed-off-by: Stefano Babic <sbabic@denx.de>