diskpart: ensure expected filesystems are present/created
Right now we only create filesystems when the partition table
changes, however this can be result in missing filesystems if
the disk already has the expected partition table layout but
does not have the expected filesystems.

Ensure that the specified filesystems are present and create
them if they are missing.

We retain the existing behavior of unconditionally creating
new filesystems in cases where the partition table is changed.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>