curl: change signature of write data callback
The "checkdwl" callback is used up now only by the Hawkbit backend to
check if there is a cancel request on the server. It is called by the
curl WRITEFUNCTION callback, but the callback in channel_curl.c changes
the signature and does not send parameters to checkdwl.

Rename checkdwl in a more generic "dwlwrdata" and pass all parameters
foreseen by Curl for the callback. This allows to get access to the
incoming stream and the channel can be used in other cases without
sending data to to installer.

Signed-off-by: Stefano Babic <sbabic@denx.de>