delta: add the option to limit size for source
When upgrading a partition, the filesystem could be much more smaller as
the partition itself. The handler as default will read the whole
partition and it will create the whole ZCK index for it, but this
requires more memory as really needed. Add a way to check the real size
of the filesystem, and do not index the rest of the partition.

Signed-off-by: Stefano Babic <sbabic@denx.de>