archive_handler: fail update on more archive errors
Do not mark update as successful if archive could not be extracted!

Current archive handler does not fail update when FS is full and cannot
be written to after archive_write_header() or archive_write_data_block().

libarchive is nice enough to return ARCHIVE_WARN when the error is not
fatal, but unfortunately inability to write a data block is returned as
a warning so the write functions treat anything non-OK as an error.

Trust warning status for read functions and only print a message in
these case as we used to do.
While we are here add entry name to error message when it makes sense.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
Acked-by: Stefano Babic <sbabic@denx.de>