util: add optional swupdate_time_iso8601 param
This adds an optional parameter to the swupdate_time_iso8601 function,
which allows to pass an older time stamp for conversion to this
function.

    char *swupdate_time_iso8601(struct timeval *tv);

The tv argument is a struct timeval (as specified in <sys/time.h>):

    struct timeval {
         time_t      tv_sec;     /* seconds since the Epoch */
         suseconds_t tv_usec;    /* microseconds */
    };

If the tv argument is NULL pointer, swupdate_time_iso8601 will obtain
the current time itself by calling gettimeofday. (Like it did before
this modification.)

Signed-off-by: Roland Gaudig <roland.gaudig@weidmueller.com>