diskpart_handler: support for GPT PARTUUIDs
It is already possible to specify PARTLABELS (using `name=`) in
SWUpdate's diskpart_handler.

This commit makes it possible to explicitly provide the PARTUUID by
setting `partuuid=` in the respective partition's property section.
If omitted, the current behavior of generating an automatic PARTUUID is
preserved, i.e. this is a backwards compatible extension.

An example use-case is to mount a partition by its PARTUUID, i.e. by
mounting /dev/disk/by-partuuid/<PARTUUID> (which is typically a symlink
created by udev). On real hardware, the symlink might point to
/dev/nvme0n1p1, whereas inside a VM, it could refer to /dev/vda1.
Thus, by using PARTUUID, the same mount command can be used in both
settings. This is useful because - unlike the disk UUID - the PARTUUID
can be explicitly set.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>