diskpart_handler: use existing uuid if not explicitely set
When the ability to specifiy a partition UUID was added it also
starting enforcing the uuid matches in is_diskpart_different,
however in cases where the uuid is not set a new one will be
randomly generated. To ensure this doesn't result in a comparison
of the partition table being marked as changed due to a different
random uuid getting generated, set the UUID based on the existing
partition if it exists in cases where one is not specified in the
sw-description.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Reviewed-by: Michael Adler <michael.adler@siemens.com>