Lua: Expose SWUpdate version to Lua handlers
With distributions picking up and packaging released
versions of SWUpdate, off-tree Lua handlers such as
swupdate-handler-roundrobin [1] have the challenge to
be compatible with the released version while also
embracing the latest development version features
(that become the next released version).

Hence, expose the SWUpdate version to the Lua realm in
order to easily implement a feature-set test instead of
having to feature-test particular functionality
individually. The base line is the current release from
which on getversion() is available to the Lua realm
after which version and patchlevel can be used to
identify a particular version's feature set.

For convenience, the returned table contains the named
fields "version" and "patchlevel" as well as "array"
fields with (in order) version, patchlevel so to allow
for (in Lua 5.1 dialect)

    version, patchlevel = unpack(swupdate.getversion())

Note that sublevel and extraversion are deliberately
ignored as they're too fine-grained for this purpose.

[1] https://gitlab.com/cip-project/cip-sw-updates/swupdate-handler-roundrobin

Signed-off-by: Christian Storm <christian.storm@siemens.com>